+++ desktop/cmp/panel/impl/Splitter.js
@@ -12,27 +12,27 @@ import {button} from '@xh/hoist/desktop/cmp/button';
 import {hbox, vbox} from '@xh/hoist/cmp/layout';
 
 import {PanelModel} from '../PanelModel';
-import './Collapser.scss';
+import './Splitter.scss';
 
 /**
  * @private
  */
 @HoistComponent
-export class Collapser extends Component {
+export class Splitter extends Component {
 
     static modelClass = PanelModel;
 
     render() {
-        const {vertical, showSplitterCollapseButton} = this.model;
+        const {vertical, showSplitterCollapseButton, collapsible} = this.model;
 
         const cmp = vertical ? hbox : vbox,
             cfg = {
-                className: `xh-resizable-collapser ${vertical ? 'vertical' : 'horizontal'}`,
+                className: `xh-resizable-splitter ${vertical ? 'vertical' : 'horizontal'}`,
                 item: button({
                     className: 'xh-resizable-collapser-btn',
                     icon: Icon[this.getChevron()](),
                     onClick: this.onClick,
-                    omit: !showSplitterCollapseButton
+                    omit: !showSplitterCollapseButton || !collapsible
                 })
             };
 
@@ -51,4 +51,4 @@ export class Collapser extends Component {
         this.model.toggleCollapsed();
     }
 }
-export const collapser = elemFactory(Collapser);
+export const splitter = elemFactory(Splitter);